+++ Control/Concurrent/STM/TSem.hs
@@ -13,7 +13,6 @@
 -- @since 2.4.2
 -----------------------------------------------------------------------------
 
-{-# LANGUAGE DeriveDataTypeable #-}
 module Control.Concurrent.STM.TSem
   ( TSem
   , newTSem
@@ -26,8 +25,7 @@ module Control.Concurrent.STM.TSem
 
 import Control.Concurrent.STM
 import Control.Monad
-import Data.Typeable
-import Data.Word as Word
+import Numeric.Natural
 
 -- | 'TSem' is a transactional semaphore.  It holds a certain number
 -- of units, and units may be acquired or released by 'waitTSem' and
@@ -44,7 +42,7 @@ import Data.Word as Word
 --
 -- @since 2.4.2
 newtype TSem = TSem (TVar Integer)
-  deriving (Eq, Typeable)
+  deriving (Eq)
 
 -- | Construct new 'TSem' with an initial counter value.
 --
@@ -56,8 +54,8 @@ newtype TSem = TSem (TVar Integer)
 -- operations to counter-balance.
 --
 -- @since 2.4.2
-newTSem :: Int -> STM TSem
-newTSem i = fmap TSem (newTVar $! toInteger i)
+newTSem :: Integer -> STM TSem
+newTSem i = fmap TSem (newTVar i)
 
 -- NOTE: we can't expose a good `TSem -> STM Int' operation as blocked
 -- 'waitTSem' aren't reliably reflected in a negative counter value.
@@ -99,7 +97,7 @@ signalTSem (TSem t) = do
 -- > signalTSem == signalTSemN 1
 --
 -- @since 2.4.5
-signalTSemN :: Word.Word -> TSem -> STM ()
+signalTSemN :: Natural -> TSem -> STM ()
 signalTSemN 0 _ = return ()
 signalTSemN 1 s = signalTSem s
 signalTSemN n (TSem t) = do